moved the internal Event Classes to own Classes in event package for easier understanding
